pyproject.toml
[project]
name = "docvqa"
version = "0.1.0"
description = "DocVQA 2026: Active Perception via RLM for Document Understanding"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.1.3",
    "datasets",
    "Pillow",
    "python-Levenshtein",
    "python-dateutil",
    "logfire[dspy,litellm]",
    "hydra-core>=1.3",
    "omegaconf",
    "python-dotenv",
    "google-cloud-aiplatform>=1.141.0",
    "bm25s>=0.3",
    "pystemmer>=3.0.0",
    "wandb>=0.25.1",
    "gepa>=0.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docvqa"]

[tool.uv]
override-dependencies = ["gepa>=0.1.1"]

[dependency-groups]
dev = [
    "huggingface-hub>=0.36.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pytest>=9.0.3",
    "wandb>=0.25.1",
]
